[postgis] Enable protobuf support
This was added as an option in #22801 but subsequently removed as part
of the general options purge.

PostGIS is not fully functional without this: certain functions,
e.g. ST_AsMVT, fail at runtime with dynamic library loading
errors. Since protobuf-c is a minor additional dependency, let's make
this the default.